Migrate remaining Planetfall god-mode commands to IGodModeCommandHandler
Follow-up to the PR-feedback refactor: the chronometer reset ("god mode
reset time|clock", from the shuttle work) and the survival-clock toggles
("god mode [no] sleep|hunger|survival", issue #277) were also game-specific
commands living in the engine's GodModeProcessor behind marker interfaces.
Move their parsing into PlanetfallContext.HandleGodModeCommand alongside
the wandering toggle, preserving the original evaluation order (reset ->
survival -> wandering), and delete the now-orphaned ISurvivalClockContext
and IResettableClockContext interfaces from Model.

The engine's built-in god-mode commands (take/go/where/kill) stay put -
they are game-agnostic, working through Repository and per-item opt-ins
(ICanBeAttacked.GodModeKill, IGodModeTeleportAware).

Behavior is unchanged; the existing end-to-end tests for all three
commands (SurvivalClockGodModeTests, ShuttleTests' reset-time cases,
UnitTests' GodModeTests) pass as-is.
